[html][canvas] Use linear gradients
You can set different gradients for stoke and fill styles.

This example assigns limegreen and white at the same position. It will
change the color sharply and the assigning order is important.

And in the stoke gradient, it starts the gradient at 0.5 instead of 0.
In this case, it will be the solid color from start to 0.5.

https://developer.mozilla.org/en-US/docs/Web/API/Canvas_API/Tutorial/Applying_styles_and_colors#a_createlineargradient_example